+++ schemas/userModel.js
@@ -1,6 +1,4 @@
 const { Schema, model } = require("mongoose");
-const Joi = require("joi");
-const bcrypt = require("bcrypt");
 
 const userSchema = Schema(
   {
@@ -35,6 +33,11 @@ const userSchema = Schema(
         "Please enter a date as 20.12.2021",
       ],
     },
+    favoriteList: {
+      type: Array,
+      ref: "notice",
+      default: [],
+    },
     token: {
       type: String,
       default: null,
@@ -43,51 +46,6 @@ const userSchema = Schema(
   { versionKey: false, timestamps: true }
 );
 
-// userSchema.pre("save", async function (next) {
-//   try {
-//     const user = this;
-//     if (!user.isModified("password")) next();
-//     const hashPassword = bcrypt.hashSync(this.password, bcrypt.genSaltSync(10));
-//     this.password = hashPassword;
-//     next();
-//   } catch (error) {
-//     return next(error);
-//   }
-// });
-// userSchema.methods.matchPassword = async function (password) {
-//   try {
-//     return await bcrypt.compare(password, this.password);
-//   } catch (error) {
-//     throw new Error(error);
-//   }
-// };
-const nameRegexp = /^[a-zA-Zа-яА-ЯіІїЇґҐ]+(?: [a-zA-Zа-яА-ЯіІїЇґҐ]+)*$/;
-const emailRegexp = /^\w+@[a-zA-Z_]+?\.[a-zA-Z]{2,3}$/;
-const passwordRegexp = /((?=.*\d)(?=.*[a-z])(?=.*[A-Z]).{7,32})/;
-const phoneRegexp = /^\+380\d{9}$/;
-const cityRegexp = /^[а-яёіїєА-ЯЁІЇЄA-Za-z]+,?\s[а-яёіїєА-ЯЁІЇЄA-Za-z]+$/;
-
-const joiUserSchema = Joi.object({
-  name: Joi.string().pattern(nameRegexp).required(),
-  password: Joi.string().pattern(passwordRegexp).min(7).max(32).required(),
-  email: Joi.string().pattern(emailRegexp).required(),
-  city: Joi.string().pattern(cityRegexp).min(2).required(),
-  phone: Joi.string().pattern(phoneRegexp).required(),
-});
-
-const joiLoginSchema = Joi.object({
-  password: Joi.string().pattern(passwordRegexp).min(7).max(32).required(),
-  email: Joi.string().pattern(emailRegexp).required(),
-});
-
-const joiUpdatedUserSchema = Joi.object({
-  name: Joi.string().pattern(nameRegexp),
-  password: Joi.string().pattern(passwordRegexp).min(7).max(32),
-  email: Joi.string().pattern(emailRegexp),
-  city: Joi.string().pattern(cityRegexp).min(2),
-  phone: Joi.string().pattern(phoneRegexp),
-});
-
 const User = model("user", userSchema);
 
-module.exports = { User, joiUserSchema, joiLoginSchema, joiUpdatedUserSchema };
+module.exports = { User };
